Rename files and structure of courses
