+++ client/src/app/components/courses/components/course-card/course-card.component.scss
