+++ client/src/app/components/courses/pages/edit/course-edit.component.scss
