+++ client/src/app/components/courses/pages/list/courses-list.component.scss
